requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.1.1
pydantic==2.5.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
pymongo==4.5.0
email-validator==2.1.0
